+++ lib/src/libedax.dart
@@ -362,9 +362,29 @@ class LibEdax {
   /// Count bit.
   int popCount(final int bit) => _bindings.bit_count(bit);
 
+  @Deprecated(
+    '''
+Use edaxBookCountBoardBestpath.
+edaxBookCountBestpath is alias of edaxBookCountBoardBestpath(board, ).
+''',
+  )
+  CountBestpathResult edaxBookCountBestpath(final Board board) {
+    final dstP = calloc<bindings.Position>();
+    final dstB = calloc<bindings.Board>();
+    dstB.ref.player = board.player;
+    dstB.ref.opponent = board.opponent;
+    _bindings.edax_book_count_bestpath(dstB, dstP);
+
+    final position = Position.fromCStruct(dstP.ref);
+    calloc
+      ..free(dstP)
+      ..free(dstB);
+    return CountBestpathResult(board, position);
+  }
+
   /// Count bestpath with book
   ///
-  /// Compute the indicator of efficiency to win, which means the minimum number you should memorize on the situation both of players always choose one of the best move list.
+  /// Compute the indicator of efficiency to win, which means the minimum number you should memorize on the situation both of players always choose one of the best move list in the range each lower limit.
   ///
   /// This function is very __advanced__. <br>
   /// __You must understand [the book structure of edax](https://choi.lavox.net/edax/book) and following important notice list__.
@@ -377,12 +397,23 @@ class LibEdax {
   /// * The depth of this feature depends on your book.
   /// * The moves which meet up with another moves is counted respectively.
   ///   * btw, symmetric moves is counted 1 because of edax book structure.
-  CountBestpathResult edaxBookCountBestpath(final Board board) {
+  CountBestpathResult edaxBookCountBoardBestpath(
+    final Board board,
+    final int playerLowerLimit,
+    final int opponentLowerLimit,
+    final int color, // return position of this color
+  ) {
     final dstP = calloc<bindings.Position>();
     final dstB = calloc<bindings.Board>();
     dstB.ref.player = board.player;
     dstB.ref.opponent = board.opponent;
-    _bindings.edax_book_count_bestpath(dstB, dstP);
+    _bindings.edax_book_count_board_bestpath(
+      dstB,
+      dstP,
+      playerLowerLimit,
+      opponentLowerLimit,
+      color,
+    );
 
     final position = Position.fromCStruct(dstP.ref);
     calloc